requirements.txt
fastapi
uvicorn 
jinja2 
pytube 
openai 
youtube_transcript_api
python-multipart
pytubefix
python-dotenv
gspread
google-auth